% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{changepoints}
\alias{changepoints}
\title{Vector of changepoints}
\usage{
changepoints(res)
}
\arguments{
\item{res}{A multiSeg object (obtained with function multiSeg)}
}
\value{
The vector of changepoints
}
\description{
Vector of changepoints
}
\examples{
n <- 500
var <- 1
means <- matrix(c(0,1,2,1,0,1,2,2,0,0,1,1),3,4)
changes <- c(0.3,0.6,1)
dataG(n,means,changes, var)
file = paste(getwd(),"/dataG.txt",sep = "")
res <- multiSeg(file,4*2*log(n),1)
changepoints(res)


}
