% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGenerator.R
\name{dataG}
\alias{dataG}
\title{Multidimensional Gaussian Data Generator}
\usage{
dataG(n, means, chgpt, sigma = 1)
}
\arguments{
\item{n}{number of obsersations to generate.}

\item{means}{matrix of means, the number of columns = number of variables = number of dimensions. The i-th raw corresponds tp the i-th segment mean in all dimensrions.}

\item{chgpt}{vector of changepoints (an increasing sequence of integers).}

\item{sigma}{standard deviation of the gaussian model}
}
\value{
The matrix of simulated data + the matrix is saved in the file dataG.txt at the current path (obtained by getwd())
}
\description{
Simulating Multidimensional Gaussian Data with a model of segmentation, that is a given matrix of means and a vector of changepoints.
}
\examples{
data <- dataG(500, matrix(c(0,1,2,1,0,1,2,2,0,0,1,1),3,4),c(0.3,0.6,1))
head(data)
}
