% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pruning}
\alias{pruning}
\title{Plots of the pruning path}
\usage{
pruning(res)
}
\arguments{
\item{res}{A multiSeg object (obtained with function multiSeg)}
}
\value{
a plot with 3 graphics. First plot = number of candidates (nb)
Second plot (in red) = delay before pruning for each candidate (delay)
Third plot = cumulative sums for all type of pruning (0, 1, 2 or 3).
type = 0 (black curve): no pruning
type = 1 (red curve): radius < 0
type = 2 (green curve): intersection = emptyset
type = 3 (blue curve): candidate hidden by one sphere of the past
}
\description{
Displaying the properties of the pruning
}
\examples{
n <- 500
var <- 1
means <- matrix(c(0,1,2,1,0,1,2,2,0,0,1,1),3,4)
changes <- c(0.3,0.6,1)
data <- dataG(n,means,changes, var)
file = paste(getwd(),"/dataG.txt",sep = "")
res <- multiSeg(file,4*2*log(n),1)
pruning(res)
res$type
}
