% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{segmentation}
\alias{segmentation}
\title{Screening of the segmentation}
\usage{
segmentation(file, res)
}
\arguments{
\item{file}{the path + name of the data file}

\item{res}{A multiSeg object (obtained with function multiSeg)}
}
\description{
Display the segments and changepoints
}
\examples{
n <- 500
var <- 1
means <- matrix(c(0,1,2,1,0,1,2,2,0,0,1,1),3,4)
changes <- c(0.3,0.6,1)
dataG(n,means,changes, var)
file = paste(getwd(),"/dataG.txt",sep = "")
res <- multiSeg(file,4*2*log(n),1)
segmentation(file,res)
}
